Resolve #13
