--- Tests/Queries/Registrations/BasicRegistrationTests.cs
@@ -1,44 +1,0 @@
-﻿using System.Threading.Tasks;
-using Xer.Cqrs.QueryStack;
-using Xer.Cqrs.QueryStack.Registrations;
-using Xer.Cqrs.Tests.Entities;
-using Xunit;
-using Xunit.Abstractions;
-
-namespace Xer.Cqrs.Tests.Queries.Registrations
-{
-    public class BasicRegistrationTests
-    {
-        #region Register Method
-        
-       public class RegisterMethod
-        {
-            private readonly ITestOutputHelper _testOutputHelper;
-
-            public RegisterMethod(ITestOutputHelper testOutputHelper)
-            {
-                _testOutputHelper = testOutputHelper;
-            }
-
-            [Fact]
-            public async Task Should_Store_All_Query_Handlers()
-            {
-                var queryHandler = new TestQueryHandler(_testOutputHelper);
-                var registration = new QueryHandlerRegistration();
-                registration.Register(() => (IQueryAsyncHandler<QuerySomething, string>)queryHandler);
-
-                QueryHandlerDelegate<string> queryHandlerDelegate = registration.ResolveQueryHandler<QuerySomething, string>();
-
-                var query = new QuerySomething(nameof(Should_Store_All_Query_Handlers));
-
-                var result1 = await queryHandler.HandleAsync(query);
-
-                var result2 = await queryHandlerDelegate.Invoke(query);
-
-                Assert.Equal(result1, result2);
-            }
-        }
-
-        #endregion Register Method
-    }
-}
