+++ Tests/Xer.Cqrs.CommandStack.Tests/Entities/Commands.cs
@@ -1,0 +1,29 @@
+﻿using Xer.Cqrs.CommandStack;
+
+namespace Xer.Cqrs.CommandStack.Tests.Entities
+{
+    public class TestCommand
+    {
+
+    }
+
+    public class ThrowExceptionCommand
+    {
+
+    }
+
+    public class CancellableTestCommand
+    {
+
+    }
+
+    public class DelayCommand
+    {
+        public int DurationInMilliSeconds { get; }
+
+        public DelayCommand(int milliSeconds)
+        {
+            DurationInMilliSeconds = milliSeconds;
+        }
+    }
+}
