--- Tests/Xer.Cqrs.Tests.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.5.0-preview-20171012-09" />
-    <PackageReference Include="MongoDB.Driver" Version="2.4.4" />
-    <PackageReference Include="SimpleInjector" Version="4.0.12" />
-    <PackageReference Include="xunit" Version="2.3.1" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.3.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Src\Xer.Cqrs.CommandStack\Xer.Cqrs.CommandStack.csproj" />
-    <ProjectReference Include="..\Src\Xer.Cqrs.EventStack\Xer.Cqrs.EventStack.csproj" />
-    <ProjectReference Include="..\Src\Xer.Cqrs.QueryStack\Xer.Cqrs.QueryStack.csproj" />
-  </ItemGroup>
-
-</Project>
