+++ Xer.Cqrs.sln
@@ -10,8 +10,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Xer.Cqrs.EventStack", "Src\
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{B8368CCD-E692-430E-AE0F-4C346C5233F5}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Xer.Cqrs.Tests", "Tests\Xer.Cqrs.Tests.csproj", "{2FFBBBB7-C2BF-4EA7-B440-3C68DD19A4CB}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{3BF5366F-3ED0-4E1B-AF2A-82E2667F4B89}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Console", "Samples\ConsoleApp\ConsoleApp.csproj", "{2C9D1FFB-07D9-4C2A-9BED-18405803FC64}"
@@ -23,6 +21,12 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReadSide", "Samples\ReadSide\ReadSide.csproj", "{3873FC0F-C1EE-4AE7-8803-1E2E32BC9D15}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure", "Samples\Infrastructure\Infrastructure.csproj", "{099D5484-CC4C-4655-AFD8-570004933044}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xer.Cqrs.CommandStack.Tests", "Tests\Xer.Cqrs.CommandStack.Tests\Xer.Cqrs.CommandStack.Tests.csproj", "{6F68B51E-810F-421F-9802-F3A9398ABAF7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xer.Cqrs.EventStack.Tests", "Tests\Xer.Cqrs.EventStack.Tests\Xer.Cqrs.EventStack.Tests.csproj", "{7B343433-0270-4DC7-BD7E-91CCBF0E4883}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xer.Cqrs.QueryStack.Tests", "Tests\Xer.Cqrs.QueryStack.Tests\Xer.Cqrs.QueryStack.Tests.csproj", "{85BC41CB-8949-49F2-884F-93BFB30AAE87}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -38,10 +42,6 @@ Global
 		{BDF57B64-636E-4C03-A7F0-986A3E9A6AF0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BDF57B64-636E-4C03-A7F0-986A3E9A6AF0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BDF57B64-636E-4C03-A7F0-986A3E9A6AF0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2FFBBBB7-C2BF-4EA7-B440-3C68DD19A4CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2FFBBBB7-C2BF-4EA7-B440-3C68DD19A4CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2FFBBBB7-C2BF-4EA7-B440-3C68DD19A4CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2FFBBBB7-C2BF-4EA7-B440-3C68DD19A4CB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3BE4E0FA-9992-4F81-BE96-9CFA9FCC1809}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3BE4E0FA-9992-4F81-BE96-9CFA9FCC1809}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3BE4E0FA-9992-4F81-BE96-9CFA9FCC1809}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -122,17 +122,55 @@ Global
 		{099D5484-CC4C-4655-AFD8-570004933044}.Release|x64.Build.0 = Release|x64
 		{099D5484-CC4C-4655-AFD8-570004933044}.Release|x86.ActiveCfg = Release|x86
 		{099D5484-CC4C-4655-AFD8-570004933044}.Release|x86.Build.0 = Release|x86
+		{6F68B51E-810F-421F-9802-F3A9398ABAF7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6F68B51E-810F-421F-9802-F3A9398ABAF7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6F68B51E-810F-421F-9802-F3A9398ABAF7}.Debug|x64.ActiveCfg = Debug|x64
+		{6F68B51E-810F-421F-9802-F3A9398ABAF7}.Debug|x64.Build.0 = Debug|x64
+		{6F68B51E-810F-421F-9802-F3A9398ABAF7}.Debug|x86.ActiveCfg = Debug|x86
+		{6F68B51E-810F-421F-9802-F3A9398ABAF7}.Debug|x86.Build.0 = Debug|x86
+		{6F68B51E-810F-421F-9802-F3A9398ABAF7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6F68B51E-810F-421F-9802-F3A9398ABAF7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6F68B51E-810F-421F-9802-F3A9398ABAF7}.Release|x64.ActiveCfg = Release|x64
+		{6F68B51E-810F-421F-9802-F3A9398ABAF7}.Release|x64.Build.0 = Release|x64
+		{6F68B51E-810F-421F-9802-F3A9398ABAF7}.Release|x86.ActiveCfg = Release|x86
+		{6F68B51E-810F-421F-9802-F3A9398ABAF7}.Release|x86.Build.0 = Release|x86
+		{7B343433-0270-4DC7-BD7E-91CCBF0E4883}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7B343433-0270-4DC7-BD7E-91CCBF0E4883}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7B343433-0270-4DC7-BD7E-91CCBF0E4883}.Debug|x64.ActiveCfg = Debug|x64
+		{7B343433-0270-4DC7-BD7E-91CCBF0E4883}.Debug|x64.Build.0 = Debug|x64
+		{7B343433-0270-4DC7-BD7E-91CCBF0E4883}.Debug|x86.ActiveCfg = Debug|x86
+		{7B343433-0270-4DC7-BD7E-91CCBF0E4883}.Debug|x86.Build.0 = Debug|x86
+		{7B343433-0270-4DC7-BD7E-91CCBF0E4883}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7B343433-0270-4DC7-BD7E-91CCBF0E4883}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7B343433-0270-4DC7-BD7E-91CCBF0E4883}.Release|x64.ActiveCfg = Release|x64
+		{7B343433-0270-4DC7-BD7E-91CCBF0E4883}.Release|x64.Build.0 = Release|x64
+		{7B343433-0270-4DC7-BD7E-91CCBF0E4883}.Release|x86.ActiveCfg = Release|x86
+		{7B343433-0270-4DC7-BD7E-91CCBF0E4883}.Release|x86.Build.0 = Release|x86
+		{85BC41CB-8949-49F2-884F-93BFB30AAE87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{85BC41CB-8949-49F2-884F-93BFB30AAE87}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{85BC41CB-8949-49F2-884F-93BFB30AAE87}.Debug|x64.ActiveCfg = Debug|x64
+		{85BC41CB-8949-49F2-884F-93BFB30AAE87}.Debug|x64.Build.0 = Debug|x64
+		{85BC41CB-8949-49F2-884F-93BFB30AAE87}.Debug|x86.ActiveCfg = Debug|x86
+		{85BC41CB-8949-49F2-884F-93BFB30AAE87}.Debug|x86.Build.0 = Debug|x86
+		{85BC41CB-8949-49F2-884F-93BFB30AAE87}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{85BC41CB-8949-49F2-884F-93BFB30AAE87}.Release|Any CPU.Build.0 = Release|Any CPU
+		{85BC41CB-8949-49F2-884F-93BFB30AAE87}.Release|x64.ActiveCfg = Release|x64
+		{85BC41CB-8949-49F2-884F-93BFB30AAE87}.Release|x64.Build.0 = Release|x64
+		{85BC41CB-8949-49F2-884F-93BFB30AAE87}.Release|x86.ActiveCfg = Release|x86
+		{85BC41CB-8949-49F2-884F-93BFB30AAE87}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{2FFBBBB7-C2BF-4EA7-B440-3C68DD19A4CB} = {B8368CCD-E692-430E-AE0F-4C346C5233F5}
 		{2C9D1FFB-07D9-4C2A-9BED-18405803FC64} = {3BF5366F-3ED0-4E1B-AF2A-82E2667F4B89}
 		{6DE1C22B-1DD2-4C71-A649-957D8295FD5C} = {3BF5366F-3ED0-4E1B-AF2A-82E2667F4B89}
 		{B1CF251D-D6C9-4C03-8112-B0F68C9ECA12} = {3BF5366F-3ED0-4E1B-AF2A-82E2667F4B89}
 		{3873FC0F-C1EE-4AE7-8803-1E2E32BC9D15} = {3BF5366F-3ED0-4E1B-AF2A-82E2667F4B89}
 		{099D5484-CC4C-4655-AFD8-570004933044} = {3BF5366F-3ED0-4E1B-AF2A-82E2667F4B89}
+		{6F68B51E-810F-421F-9802-F3A9398ABAF7} = {B8368CCD-E692-430E-AE0F-4C346C5233F5}
+		{7B343433-0270-4DC7-BD7E-91CCBF0E4883} = {B8368CCD-E692-430E-AE0F-4C346C5233F5}
+		{85BC41CB-8949-49F2-884F-93BFB30AAE87} = {B8368CCD-E692-430E-AE0F-4C346C5233F5}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {03714871-4DA2-4441-880E-E509A0F8205E}
